Implemented CRUD operations in CSV format for the majority of the app's functionality. Admins can now delete users.

+++ src/confectionery/Model/Cake.java
@@ -12,7 +12,7 @@ public class Cake extends Product implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
     private int calories;
-    public Cake(int idProduct, String name, float price, float weight, ExpirationDate expirationDate,int points, int calories) {
+    public Cake(int idProduct, String name, double price, double weight, ExpirationDate expirationDate,int points, int calories) {
         super(idProduct, name, price, weight, expirationDate,points);
         this.calories=calories;
     }
